+++ src/utils/tools.hpp
@@ -9,6 +9,7 @@
 
 #include <string>
 #include <fstream>
+#include <vector>
 
 namespace tools {
     /**
@@ -36,35 +37,22 @@ namespace tools {
         return path.substr(path.find_last_of('.') + 1);
     }
 
-    std::vector<std::string> split(const std::string &src, const char delim = ',') {
-        std::vector<std::string> res;
+    /**
+     * Returns keys of the map
+     *
+     * @tparam K Type of the key
+     * @tparam V Type of the value
+     * @param input Map to get keys
+     * @return Keys of the map
+     */
+    template<typename K, typename V>
+    std::vector<K> keys(const std::map<K, V> &input) {
+        std::vector<K> keys;
+        for (auto &it: input)
+            keys.push_back(it.first);
 
-        std::string current;
-        for (const auto c: src) {
-            if (c == delim) {
-                res.push_back(current);
-                current = "";
-            } else {
-                current += c;
-            }
-        }
-
-        if (!current.empty())
-            res.push_back(current);
-
-        return res;
+        return keys;
     }
-
-    // ====================
-    // Console colors
-
-    const std::string reset = "\033[0m";
-
-    const std::string red = "\033[31m";
-
-    const std::string magenta = "\033[35m";
-
-    const std::string bold = "\033[1m";
 }
 
 #endif //AVOW_TOOLS_HPP